odometer + float +  expenses filter

+++ Hurricane/CoreData/Vehicle+CoreDataProperties.swift
@@ -23,7 +23,7 @@ extension Vehicle {
     @NSManaged public var fuelTypeTwo: Int32
     @NSManaged public var model: String?
     @NSManaged public var name: String?
-    @NSManaged public var odometer: Int32
+    @NSManaged public var odometer: Float
     @NSManaged public var plate: String?
     @NSManaged public var year: Int32
     @NSManaged public var expenses: NSSet?
